Add files via upload
Upload initial public version of an excel / libreoffice table, which uses some yearly germany power grid data from https://www.energy-charts.info, and allows to extrapolate how it would look like, with different amounts of installed renewable energy or storage.
Play with the input data fields, like the factors, which are applied then to the records of the selected yearly recorded energy data.

The data from energy charts was downloaded and imported as .csv into the table and contains 15 minute intervals of power grid details. This data is reduced to a set of hourly data, for which the data are then modified based on the input and processed.